Allow disabled=true even when submitting
When implementing auto-saving TextFields it is a
problem when the field becomes disabled when submitting
the form. Focus is lost on the field, interrupting
the user's entry.

The existing logic only allows for setting the disabled
prop to true, but if set to false it always clobbers it
with the value of isSubmitting. This change always
respects the value of the disabled prop if set,
regardless of the value.